refactor(core): move prepareForJsonOutput from cli to core layer (#24)
Importing from cli.ts caused program.parse() to fire in test
environments, triggering process.exit(2) on CI. Extract
prepareForJsonOutput to core/prepare-output.ts where it belongs
architecturally (core = wiring layer, cli = thin shell).

Also resolves #25 and the review finding about adapter tests
importing from the CLI layer.

Closes #25
